tests: move key setup to test-utils
